Refactor application related component names
